Split off log level settings into separate files
