+++ src/Core/Settings/LoggingSettings/AdminLogLevelSettings.cs
@@ -1,0 +1,8 @@
+using Serilog.Events;
+
+namespace Bit.Core.Settings.LogSettings;
+
+public class AdminLogLevelSettings : IAdminLogLevelSettings
+{
+    public LogEventLevel Default { get; set; } = LogEventLevel.Error;
+}
