+++ src/Core/Settings/LoggingSettings/EventsLogLevelSettings.cs
@@ -1,0 +1,9 @@
+using Serilog.Events;
+
+namespace Bit.Core.Settings.LogSettings;
+
+public class EventsLogLevelSettings : IEventsLogLevelSettings
+{
+    public LogEventLevel Default { get; set; } = LogEventLevel.Error;
+    public LogEventLevel IdentityToken { get; set; } = LogEventLevel.Fatal;
+}
