+++ src/Core/Settings/LoggingSettings/EventsProcessorLogLevelSettings.cs
@@ -1,0 +1,8 @@
+using Serilog.Events;
+
+namespace Bit.Core.Settings.LogSettings;
+
+public class EventsProcessorLogLevelSettings : IEventsProcessorLogLevelSettings
+{
+    public LogEventLevel Default { get; set; } = LogEventLevel.Warning;
+}
