+++ src/Core/Settings/LoggingSettings/IconsLogLevelSettings.cs
@@ -1,0 +1,8 @@
+using Serilog.Events;
+
+namespace Bit.Core.Settings.LogSettings;
+
+public class IconsLogLevelSettings : IIconsLogLevelSettings
+{
+    public LogEventLevel Default { get; set; } = LogEventLevel.Error;
+}
