+++ src/Core/Settings/LoggingSettings/IdentityLogLevelSettings.cs
@@ -1,0 +1,10 @@
+using Serilog.Events;
+
+namespace Bit.Core.Settings.LogSettings;
+
+public class IdentityLogLevelSettings : IIdentityLogLevelSettings
+{
+    public LogEventLevel Default { get; set; } = LogEventLevel.Error;
+    public LogEventLevel IdentityToken { get; set; } = LogEventLevel.Fatal;
+    public LogEventLevel IpRateLimit { get; set; } = LogEventLevel.Information;
+}
