+++ src/Core/Settings/LoggingSettings/NotificationsLogLevelSettings.cs
@@ -1,0 +1,9 @@
+using Serilog.Events;
+
+namespace Bit.Core.Settings.LogSettings;
+
+public class NotificationsLogLevelSettings : INotificationsLogLevelSettings
+{
+    public LogEventLevel Default { get; set; } = LogEventLevel.Warning;
+    public LogEventLevel IdentityToken { get; set; } = LogEventLevel.Fatal;
+}
