+++ src/Core/Settings/LoggingSettings/SsoLogLevelSettings.cs
@@ -1,0 +1,8 @@
+using Serilog.Events;
+
+namespace Bit.Core.Settings.LogSettings;
+
+public class SsoLogLevelSettings : ISsoLogLevelSettings
+{
+    public LogEventLevel Default { get; set; } = LogEventLevel.Error;
+}
